# bump rust-bison-skeleton; document new performance vs ripper