--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,7 +18,7 @@ onig = {version = "6", optional = true}
 
 [build-dependencies]
 lib-ruby-parser-nodes = {version = "0.7.0", optional = true}
-rust-bison-skeleton = {version = "0.11.0", optional = true}
+rust-bison-skeleton = {version = "0.12.0", optional = true}
 
 [dev-dependencies]
 clap = "3.0.0-beta.2"