--- a/src/parse_value.rs
+++ b/src/parse_value.rs
@@ -621,3 +621,9 @@ impl ParseValue {
         Self::MatchPatternWithTrailingComma(Box::new(value))
     }
 }
+
+impl Default for ParseValue {
+    fn default() -> Self {
+        Self::Stolen
+    }
+}