New version of `rust-bison-skeleton` does way less allocations in `parse` method. Now `lib-ruby-parser` significantly outperforms Ripper:

```
repeat 10 time target/release/examples/parse --no-output 'gems/repos/**/*.rb'; repeat 10 time ruby -rripper -e "Dir['gems/repos/**/*.rb'].each { |f| Ripper.sexp(File.read(f)) }"
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  9.52s user 1.22s system 81% cpu 13.158 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.64s user 0.67s system 99% cpu 9.315 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.66s user 0.65s system 99% cpu 9.311 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.61s user 0.63s system 99% cpu 9.251 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.83s user 0.65s system 99% cpu 9.492 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.75s user 0.65s system 99% cpu 9.411 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.71s user 0.64s system 99% cpu 9.353 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.67s user 0.64s system 99% cpu 9.317 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.64s user 0.65s system 99% cpu 9.296 total
Done
target/release/examples/parse --no-output 'gems/repos/**/*.rb'  8.62s user 0.63s system 99% cpu 9.253 total
ruby -rripper -e   16.37s user 0.94s system 99% cpu 17.314 total
ruby -rripper -e   16.37s user 0.90s system 99% cpu 17.273 total
ruby -rripper -e   16.17s user 0.93s system 99% cpu 17.108 total
ruby -rripper -e   16.20s user 0.93s system 99% cpu 17.140 total
ruby -rripper -e   16.10s user 0.93s system 99% cpu 17.028 total
ruby -rripper -e   16.46s user 0.95s system 99% cpu 17.412 total
ruby -rripper -e   16.23s user 0.93s system 99% cpu 17.190 total
ruby -rripper -e   16.35s user 0.96s system 99% cpu 17.324 total
ruby -rripper -e   16.27s user 0.94s system 99% cpu 17.223 total
ruby -rripper -e   16.42s user 0.95s system 99% cpu 17.379 total
```